requirements-gcp.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.1
defusedxml==0.7.1
Django==4.2.1
django-cleanup==8.0.0
django-environ==0.11.2
django-extra-views==0.14.0
django-searchable-encrypted-fields==0.2.1
django-storages==1.14.2
exceptiongroup==1.1.3
fonttools==4.43.1
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-resources==6.1.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.327
langsmith==0.0.56
marshmallow==3.20.1
matplotlib==3.8.1
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
numpy==1.26.1
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
plotly==5.18.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
sorl-thumbnail==12.10.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
tenacity==8.2.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.7
webencodings==0.5.1
wordcloud==1.9.2
yarl==1.9.2
zipp==3.17.0
